+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "sestat"
-version = "1.1.0"
+version = "1.0.2"
 authors = [
   { name="Dainius Kirsnauskas", email="dainius@genomika.lt" },
 ]
